+++ src/civilization/concepts/trade/mod.rs
@@ -4,3 +4,4 @@ pub mod triggers;
 pub mod events;
 pub mod plugins;
 pub mod components;
+pub mod functions;
